client/requirements.txt
pygame
requests
python-socketio
